continual_learning_ids/requirements.txt
# ══════════════════════════════════════════════════════════════════════
# Requirements for CL-RL NIDS Framework
# Continual Learning and Constrained RL for Adversarially Robust
# Network Intrusion Detection and Autonomous Response
# ══════════════════════════════════════════════════════════════════════

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Logging & Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Kaggle dataset download
# kaggle>=1.5.0
